requirements/requirements.txt
sanic
python-socketio
aiohttp
pyyaml
sanic_jinja2
sanic_session
aioorm==0.1.2
requests_html
sanic_mail
aioredis
ua-parser 
user-agents
itsdangerous